修复开卡模式下错误移除歌榜

+++ modules/ranking.php
@@ -26,7 +26,7 @@ function ranking_live($post) {
   while ($item = $rank->fetch()) {
     $ret2['rank'] = (int)$item['rank'];
     $ret2['score'] = (int)$item['hi_score'];
-    if($ret2['score']>$score_max){
+    if($params['card_switch']==0 && $ret2['score']>$score_max){
       $mysql->query("DELETE FROM live_ranking 
       WHERE user_id=".$item['user_id']."
       AND notes_setting_asset='".$notes_setting['notes_setting_asset']."'
